apps/ml-pipeline/requirements.txt
osmium
psycopg2-binary
shapely
sqlalchemy
python-dotenv
earthengine-api
xgboost
scikit-learn
numpy
httpx
